+++ .gitignore
@@ -40,6 +40,7 @@ build/Release
 
 # Dependency directories
 node_modules/
+src/node_modules/
 jspm_packages/
 
 # Snowpack dependency directory (https://snowpack.dev/)
