+++ README.md
@@ -1,1 +1,39 @@
-# gharPaluwa
+# Gharpaluwa Market Place
+
+## Description
+
+Gharpaluwa Market Place is a web application that serves as an online marketplace for buying and selling various products. Users can browse through a wide range of products, add items to their cart, and make purchases securely. The application provides a convenient platform for both buyers and sellers to interact and conduct transactions.
+
+🛒 Browse and purchase a variety of products
+💼 Convenient platform for buyers and sellers
+🔒 Secure transactions and payment processing
+
+## Installation
+
+To install the required packages, run the following command in the terminal:
+
+```bash
+npm install
+```
+
+## Running the Application
+
+To start the application, use the following command:
+
+```bash
+npm run dev
+```
+
+## Nodemon Installation
+
+If you haven't installed nodemon globally on your device, you can use the following command:
+
+```bash
+npm install -g nodemon
+```
+
+Please note that nodemon is used to automatically restart the server whenever a file is modified, which is useful during development.
+
+## License
+
+This project is licensed under the [MIT License](LICENSE). Feel free to modify and distribute it as per the license terms.
