--- controllers/index.js
@@ -1,7 +1,0 @@
-import {login,register} from "./auth/auth";
-import { getitem } from "./manage/posts/post";
-
-
-export {
-    login,register,getitem
-}
