--- middleware/index.js
@@ -1,3 +1,0 @@
-import { authenticateToken } from "./auth/authenticatetoken";
-
-export {authenticateToken};
