+++ package.json
@@ -2,11 +2,11 @@
   "name": "gharpaluwa",
   "version": "1.0.0",
   "description": "",
-  "main": "app.js",
-  "type":"module",
+  "main": "src/app.js",
+  "type": "module",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "dev": "nodemon app.js "
+    "dev": "nodemon src/app.js "
   },
   "keywords": [],
   "author": "",
@@ -16,7 +16,9 @@
     "dotenv": "^16.0.3",
     "express": "^4.18.2",
     "jsonwebtoken": "^9.0.0",
-    "mongoose": "^7.1.0",
+    "mongoose": "^7.1.0"
+  },
+  "devDependencies": {
     "nodemon": "^2.0.22"
   }
 }
