+++ src/controllers/index.js
@@ -1,0 +1,6 @@
+import { login, register } from "./auth/authController.js";
+import { getItem } from "./manage/items/itemController.js";
+
+// =========Importing and Exporting All Controllers================
+// =========Controllers are the functions that are called when a route is hit.================
+export { login, register, getItem };
