+++ src/database/connection.js
@@ -1,0 +1,17 @@
+import { connect } from "mongoose";
+
+const connectDb = (url) => {
+  connect(url, {
+    useNewUrlParser: true,
+    useUnifiedTopology: true,
+  })
+    .then(() => {
+      console.log("connection sucessfull");
+    })
+    .catch((e) => {
+      console.log("not connected");
+      console.log(e);
+    });
+};
+
+export default connectDb;
