+++ src/middleware/index.js
@@ -1,0 +1,3 @@
+import { authenticateToken } from "./auth/authenticatetoken.js";
+
+export { authenticateToken };
